move playback to won component
