+++ src/timebar/timebar.module.css
@@ -30,10 +30,6 @@
   cursor: not-allowed;
 }
 
-.playbackActions svg {
-  fill: var(--timebar-white);
-}
-
 .uiButton:disabled svg {
   opacity: 0.5;
 }
@@ -42,45 +38,6 @@
 .uiButton:focus {
   background-color: var(--timebar-ui-btn-color-hover);
   outline: none;
-}
-
-/* PLAYBACK ACTIONS */
-
-.playbackActions {
-  background-color: var(--timebar-ui-btn-color);
-  width: var(--timebar-size);
-  height: 100%;
-  margin-right: 0.1rem;
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.secondary {
-  opacity: 0;
-}
-
-.playbackActionsActive .secondary,
-.playbackActions:hover .secondary {
-  opacity: 0.5;
-}
-
-.playbackActions .secondaryActive,
-.playbackActions:hover .secondary:hover {
-  opacity: 1;
-}
-
-.playbackActions .uiButton {
-  width: 2.4rem;
-  height: 2.4rem;
-  border-radius: 50%;
-}
-
-.playbackActions .loop {
-  margin: 0 2.4rem;
-}
-
-.playbackActions .speed {
-  margin-left: 2.4rem;
 }
 
 /* TIME ACTIONS */
@@ -105,6 +62,11 @@
   border-top: 0.1rem solid var(--timebar-main-background-color);
   height: 3.5rem;
 }
+
+.timeRangeContainer {
+  position: relative;
+}
+
 
 @media only screen and (max-width: 72rem) {
   :root {
